requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
google-generativeai==0.3.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.39
typing_extensions==4.12.2
uuid==1.30
Werkzeug==3.1.3
